refactor(postgres): extract partitioning into its own module
Move PostgresPartitionConfig and the range-bound helpers/enums out of
impl.py into dbt/adapters/postgres/partitioning.py. impl.py imports the
config and a compute_partition_bounds() helper; the @available
get_partition_bounds method now just delegates.
